+++ src/operation-menu/OperationMenuList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Menu, { MenuItem } from '@gio-design/components/es/components/menu';
-import Tooltip from '@gio-design/components/es/components/tooltip';
+import { Tooltip } from '@gio-design/components';
 import { OperationMenuListOption, OperationMenuListProps } from './interfaces';
 
 const OperationMenuList = (props: OperationMenuListProps) => {
@@ -14,7 +14,7 @@ const OperationMenuList = (props: OperationMenuListProps) => {
           .map((option: OperationMenuListOption) => (
             <MenuItem icon={option.icon} key={option.value} disabled={option.disabled}>
               <Tooltip title={option.disabled ? option.tooltip : ''} placement="leftTop">
-                <>{option.label}</>
+                <span style={{ display: 'block', marginLeft: -20, paddingLeft: 20 }}>{option.label}</span>
               </Tooltip>
             </MenuItem>
           ))}